requirements.txt
flask
flask-cors
Flask-JWT
flask-restful
Flask-SQLAlchemy
PyMySQL
mysqlclient
python-dotenv
